refactor: extract landing page logic into custom hooks to reduce code duplication
